Conflicts:
	WaywardLib/src/main/java/net/wayward_realms/waywardlib/monsters/MonstersPlugin.java
	WaywardMonsters/src/main/java/net/wayward_realms/waywardmonsters/EntityLevelManager.java
	WaywardMonsters/src/main/java/net/wayward_realms/waywardmonsters/WaywardMonsters.java
	WaywardMonsters/src/main/resources/plugin.yml

+++ WaywardLib/src/main/java/net/wayward_realms/waywardlib/events/EventsPlugin.java
@@ -47,7 +47,7 @@ public interface EventsPlugin extends WaywardPlugin {
      * @param id the id
      * @return the event character if there is one by the ID, otherwise null if the character doesn't exist, or is a normal character
      */
-    public EventCharacter getEventCharacter(long id);
+    public EventCharacter getEventCharacter(int id);
 
     /**
      * Creates a new event character for the given player
